requirements.txt
django >=4.1.4
djangorestframework
django-cors-headers
beautifulsoup4
lxml
requests
django-cleanup
Pillow
pyyaml
django-rest-swagger
